backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
openai
pydantic-settings
openpyxl
requests
reportlab
